Template Method v1.0
The  basic process all algorithms should follow, so everyone can focus on the actual algorithm's implementation.
the String[][]s used by this template may be replaced depending on how we finalize handling of our data sets. 
The very generic print statement at the conclusion of the run method will likely be refined.
The run method requires a filled data set in a 2d array as input to operate.